projects/abc_application/app03/requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
psutil==5.8.0
psycopg2-binary==2.9.3